Modify User registration

+++ src/utils/cloudinary.js
@@ -10,7 +10,9 @@ cloudinary.config({
 
 const uploadOnCloudinary = async(localFilePath) => {
     try {
-        if(!localFilePath) return null
+        if(!localFilePath){
+            return null;
+        } 
 
         // Uplaod the file on cloudinary
         const response = await cloudinary.uploader.upload(localFilePath, {
@@ -19,11 +21,15 @@ const uploadOnCloudinary = async(localFilePath) => {
 
         // File has been uploaded successfully
         // console.log("File is uploaded on cloudinary", response.url)
-        fs.unlink(localFilePath)
+        fs.unlinkSync(localFilePath)
         return response;
+
     } catch (error) {
-        fs.unlinkSync(localFilePath) //Remove the locally saved temporary file as the upload operation got failed
-        return null
+        if (localFilePath && fs.existsSync(localFilePath)){
+            fs.unlinkSync(localFilePath) //Remove the locally saved temporary file as the upload operation got failed
+            return null;
+        }
+        
     }
 }
